requirements.txt
supabase==2.10.0
python-dotenv==1.0.1
pypdf==5.1.0
langchain==0.3.7
langchain-openai==0.2.10
pydantic==2.9.2